pyproject.toml
[project]
name = "zigbee2mqtt-health"
version = "0.1.0"
description = "Monitor Zigbee2MQTT devices and publish health status to MQTT."
authors = [
    {name = "Sheya Bernstein",email = "sheyaberns@gmail.com"}
]
license = "MIT"
requires-python = ">=3.12,<4.0"

dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "backoff (>=2.2.1,<3.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "ruff (>=0.14.1,<0.15.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "pytest-env (>=1.2.0,<2.0.0)"
]

[tool.black]
line-length = 120
target-version = ['py314']

[tool.ruff]
line-length = 120
lint.select = ["E", "F", "I", "C", "B"]
lint.ignore = ["E203"]

[tool.pytest.ini_options]
addopts = "-v --cov=zigbee2mqtt_health --cov-report=term-missing"

[tool.pytest_env]
MQTT_BROKER = "localhost"
MQTT_PORT = 1883
MQTT_USERNAME = "user"
MQTT_PASSWORD = "password"
LOG_LEVEL = "debug"
